Move workers orchestration package to internal/workerctl
The root workers package (build.go, log.go, run.go) is application logic for
building and running worker subprocesses, with no consumer outside the root
module. Application code that isn't part of a public API belongs under
internal/, where the import boundary is compiler-enforced. It also could not
stay at workers/, since that directory is the documented polyglot
worker-source home (workers/<lang>/...).

Move it to internal/workerctl (package workerctl) and update importers. The
workers/ directory now contains only the per-language worker source trees.

Testing: `go build ./...` and `go vet ./...` pass for both modules.

stack-info: PR: https://github.com/temporalio/omes/pull/373, branch: lilydoar/stack/7
