+++ internal/workerctl/build.go
@@ -1,4 +1,4 @@
-package workers
+package workerctl
 
 import (
 	"context"
